backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
pytest
httpx
google-generativeai
sqlalchemy
aiosqlite
